Refactor trotter module: replace TrotterCircuit class with function-based API
Rename trotter.py to trotterization.py and replace the class-based design
with a standalone trotterized_time_evolution() function. The Hamiltonian
decomposition responsibility is moved to the caller; this module now accepts
pre-split observable lists only.
