+++ qamomile/circuit/algorithm/__init__.py
@@ -27,7 +27,10 @@ from .qaoa import (
     qaoa_state,
     x_mixer,
 )
-from .trotter import TrotterCircuit
+from .trotterization import (
+    product_formula,
+    trotterized_time_evolution,
+)
 
 __all__ = [
     # QAOA
@@ -54,6 +57,7 @@ __all__ = [
     "cost_layer",
     "fqaoa_layers",
     "fqaoa_state",
-    # Trotter
-    "TrotterCircuit",
+    # Trotterization
+    "trotterized_time_evolution",
+    "product_formula",
 ]
